refactor event key translation into key KeyMap module so all key descriptor normalization is in one place
